requirements.txt
fastapi
uvicorn
numpy
opencv-python
python-multipart
pydantic
python-dotenv
torch
torchvision
# torch --index-url https://download.pytorch.org/whl/cpu
# torchvision --index-url https://download.pytorch.org/whl/cpu
pillow
scikit-learn
celery
redis
huggingface_hub
requests
gunicorn
boto3